Develop

+++ app/Http/Controllers/Admin/ProductController.php
@@ -101,10 +101,10 @@ class ProductController extends Controller
         $image = $product->images->first();
 
 
-        if (Storage::disk('public')->exists($image->path)) {
+        if ($image and Storage::disk('public')->exists($image->path)) {
             Storage::disk('public')->delete($image->path);
+            $image->delete();
         }
-        $image->delete();
         $product->delete();
 
         return back()->with('success', 'Producto eliminado correctamente');
